+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - SQL Server create database script in db/ folder
 - Documentation in docs/ folder to update the manual for version 4
 - Metadata description and structured data for Events
+- Trigger activation date
 
 ### Fixed
 - Broken household URLs in Mission Control (leading spaces)